Updated ADC read functions to return type uint16_t to avoid confusions with integer signedness

+++ STM32F0-Utilities/STM32F0-Utilities/include/STM32F0_ADC.h
@@ -44,5 +44,5 @@ Functions to use the analog to digital converter
 /* FUNCTIONS */
 
 void init_ADC(int resolution); // Initialise and calibrate the ADC
-int analogRead(IOPin_TypeDef* iopin); // Read an analog value from a pin
-int analogReadChannel(int channel); // Read an analog value from a specific ADC channel
+uint16_t analogRead(IOPin_TypeDef* iopin); // Read an analog value from a pin
+uint16_t analogReadChannel(int channel); // Read an analog value from a specific ADC channel
